Added unit test. Add integration test.
