+++ tests/unit_tests/utils/test_utils.py
@@ -15,11 +15,11 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-from bittensor_wallet import Wallet, Keyfile, Keypair
+from bittensor_wallet import Wallet
+import pytest
 
 from bittensor import utils
 from bittensor.core.settings import SS58_FORMAT
-import pytest
 
 
 def test_ss58_to_vec_u8(mocker):
@@ -171,26 +171,28 @@ def test_is_valid_bittensor_address_or_public_key(mocker, test_input, expected_r
     assert result == expected_result
 
 
-def test_unlock_key(monkeypatch):
-    wallet = Wallet(path="/tmp/wallets")
-    cold_kf = Keyfile("/tmp/wallets/default/coldkey", name="default")
-    kp = Keypair.create_from_mnemonic(
-        "stool feel open east woman high can denial forget screen trust salt"
-    )
-    cold_kf.set_keypair(kp, False, False)
-    cold_kf.encrypt("1234password1234")
-    hot_kf = Keyfile("/tmp/wallets/default/hotkey", name="default")
-    hkp = Keypair.create_from_mnemonic(
-        "stool feel open east woman high can denial forget screen trust salt"
-    )
-    hot_kf.set_keypair(hkp, False, False)
-    hot_kf.encrypt("1234hotkey1234")
-    monkeypatch.setattr("getpass.getpass", lambda _: "badpassword1234")
-    result = utils.unlock_key(wallet)
-    assert result.success is False
-    monkeypatch.setattr("getpass.getpass", lambda _: "1234password1234")
-    result = utils.unlock_key(wallet)
-    assert result.success is True
-    monkeypatch.setattr("getpass.getpass", lambda _: "badpassword1234")
-    result = utils.unlock_key(wallet, "hot")
-    assert result.success is False
+@pytest.mark.parametrize(
+    "unlock_type, wallet_method",
+    [
+        ("cold", "unlock_coldkey"),
+        ("hot", "unlock_hotkey"),
+        ("coldkeypub", "unlock_coldkeypub"),
+    ],
+)
+def test_unlock_key(mocker, unlock_type, wallet_method):
+    """Test the unlock key function."""
+    # Preps
+    mock_wallet = mocker.Mock(autospec=Wallet)
+
+    # Call
+    result = utils.unlock_key(mock_wallet, unlock_type=unlock_type)
+
+    # Asserts
+    getattr(mock_wallet, wallet_method).assert_called_once()
+    assert result == utils.UnlockStatus(True, "")
+
+
+def test_unlock_key_raise_value_error(mocker):
+    """Test the unlock key function raises ValueError."""
+    with pytest.raises(ValueError):
+        utils.unlock_key(wallet=mocker.Mock(autospec=Wallet), unlock_type="coldkey")
